Added to transfer methods client-side and server-side

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Transfer.java
@@ -24,7 +24,7 @@ public class Transfer {
     private Integer accountToId;
     @NotEmpty
     @JsonProperty("amount")
-    private BigDecimal amount;
+    private double amount;
 
     public Transfer() {
     }
@@ -69,11 +69,11 @@ public class Transfer {
         this.accountToId = accountToId;
     }
 
-    public BigDecimal getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 }
